requirements-recsys.txt
sagemaker
boto3
pandas
pyarrow
numpy
pyyaml
scikit-learn
